Fix #1349630: This fixes the etl side of the synfig amd64 FTBFS
git-svn-id: http://svn.voria.com/code@154 1f10aa63-cdf2-0310-b900-c93c546f37ac

+++ ETL/trunk/ETL/_surface.h
@@ -274,8 +274,8 @@ public:
 		}
 				
 		//clip width against dest width
-		w = std::min(w,DEST_PEN.end_x()-DEST_PEN.x());
-		h = std::min(h,DEST_PEN.end_y()-DEST_PEN.y());
+		w = std::min(w,(int)(DEST_PEN.end_x()-DEST_PEN.x()));
+		h = std::min(h,(int)(DEST_PEN.end_y()-DEST_PEN.y()));
 		
 		//clip width against src width
 		w = std::min(w,w_-x);
